[MOV] project_timesheet: extract timesheet on task from hr_timesheet
Impacted modules: hr_timesheet, project_timesheet,
sale_timesheet

Since the previous commit change the semantic of a timesheet
line, it is now allow to create timesheet without project.
This commit made it effective by moving all code related to
"timesheet on project" into the reborn project_timesheet
module.

We now have
- hr_timesheet: responsible of the 'is_timesheet' flag, and
allowing to create timesheet
- project_timesheet: allowing to create timesheet on tasks

nothing really change functionnal except that installing
hr_timesheet (not an app) will allow user to log timesheets
on an analytic account (and not a project).

Billing timesheet flow does not change, we still need to
have go through project/task to increase delivery quanity
on sale lines. This change will come on the next commit.
This is why sale_timeshett temporary depends on
project_timesheet.

Task-1923949
