+++ addons/hr_timesheet/data/hr_timesheet_data.xml
@@ -1,13 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
 
-	<!-- For a perfect onboarding, we want the tour of project to be consumed
-	only by admin when installing hr_timesheet, to avoid confusing -->
-    <record id="web_tour_project_consumed_by_admin" model="web_tour.tour">
-        <field name="name">project_tour</field>
-        <field name="user_id" ref='base.user_admin'/>
-    </record>
-
     <!-- Set the JS widget -->
     <record id="uom.product_uom_day" model="uom.uom">
         <field name="timesheet_widget">float_toggle</field>
@@ -16,11 +9,5 @@
     <record id="uom.product_uom_hour" model="uom.uom">
         <field name="timesheet_widget">float_time</field>
     </record>
-
-    <!-- Force Analytic account creation for projects allowing timesheet (default is True) -->
-    <function
-        model="project.project"
-        name="_init_data_analytic_account"
-        eval="[]"/>
 
 </odoo>
