+++ addons/hr_timesheet/models/__init__.py
@@ -1,11 +1,9 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import analytic_account
 from . import hr
 from . import hr_timesheet
 from . import ir_http
 from . import res_company
 from . import res_config_settings
-from . import project
 from . import uom
