--- addons/hr_timesheet/report/__init__.py
@@ -1,1 +1,0 @@
-from . import project_report
