+++ addons/hr_timesheet/report/hr_timesheet_report_view.xml
@@ -39,82 +39,6 @@
             <field name="act_window_id" ref="act_hr_timesheet_report"/>
         </record>
 
-        <record id="timesheet_action_report_by_project" model="ir.actions.act_window">
-            <field name="name">By Project</field>
-            <field name="res_model">account.analytic.line</field>
-            <field name="domain">[('is_timesheet', '=', 'True')]</field>
-            <field name="context">{'search_default_groupby_project':1,}</field>
-            <field name="search_view_id" ref="hr_timesheet_line_search"/>
-            <field name="help" type="html">
-              <p class="o_view_nocontent_smiling_face">
-                Record a new activity
-              </p><p>
-                You can register and track your workings hours by project every
-                day. Every time spent on a project will become a cost and can be re-invoiced to
-                customers if required.
-              </p>
-            </field>
-        </record>
-
-        <record id="timesheet_action_view_report_by_project_pivot" model="ir.actions.act_window.view">
-            <field name="sequence" eval="5"/>
-            <field name="view_mode">pivot</field>
-            <field name="view_id" ref="hr_timesheet.view_hr_timesheet_line_pivot"/>
-            <field name="act_window_id" ref="timesheet_action_report_by_project"/>
-        </record>
-
-        <record id="timesheet_action_view_report_by_project_tree" model="ir.actions.act_window.view">
-            <field name="sequence" eval="10"/>
-            <field name="view_mode">tree</field>
-            <field name="view_id" ref="hr_timesheet.hr_timesheet_line_tree"/>
-            <field name="act_window_id" ref="timesheet_action_report_by_project"/>
-        </record>
-
-        <record id="timesheet_action_view_report_by_project_form" model="ir.actions.act_window.view">
-            <field name="sequence" eval="15"/>
-            <field name="view_mode">form</field>
-            <field name="view_id" ref="hr_timesheet.hr_timesheet_line_form"/>
-            <field name="act_window_id" ref="timesheet_action_report_by_project"/>
-        </record>
-
-        <record id="timesheet_action_report_by_task" model="ir.actions.act_window">
-            <field name="name">By Task</field>
-            <field name="res_model">account.analytic.line</field>
-            <field name="domain">[('is_timesheet', '=', 'True')]</field>
-            <field name="context">{'search_default_groupby_task':1,}</field>
-            <field name="search_view_id" ref="hr_timesheet_line_search"/>
-            <field name="help" type="html">
-              <p class="o_view_nocontent_smiling_face">
-                Record a new activity
-              </p><p>
-                You can register and track your workings hours by project every
-                day. Every time spent on a project will become a cost and can be re-invoiced to
-                customers if required.
-              </p>
-            </field>
-        </record>
-
-        <record id="timesheet_action_view_report_by_task_pivot" model="ir.actions.act_window.view">
-            <field name="sequence" eval="5"/>
-            <field name="view_mode">pivot</field>
-            <field name="view_id" ref="hr_timesheet.view_hr_timesheet_line_pivot"/>
-            <field name="act_window_id" ref="timesheet_action_report_by_task"/>
-        </record>
-
-        <record id="timesheet_action_view_report_by_task_tree" model="ir.actions.act_window.view">
-            <field name="sequence" eval="10"/>
-            <field name="view_mode">tree</field>
-            <field name="view_id" ref="hr_timesheet.hr_timesheet_line_tree"/>
-            <field name="act_window_id" ref="timesheet_action_report_by_task"/>
-        </record>
-
-        <record id="timesheet_action_view_report_by_task_form" model="ir.actions.act_window.view">
-            <field name="sequence" eval="15"/>
-            <field name="view_mode">form</field>
-            <field name="view_id" ref="hr_timesheet.hr_timesheet_line_form"/>
-            <field name="act_window_id" ref="timesheet_action_report_by_task"/>
-        </record>
-
         <menuitem id="menu_timesheets_reports"
             name="Reporting"
             parent="timesheet_menu_root"
@@ -132,18 +56,6 @@
             action="act_hr_timesheet_report"
             name="By Employee"
             sequence="10"/>
-
-        <menuitem id="timesheet_menu_report_timesheet_by_project"
-            parent="menu_timesheets_reports_timesheet"
-            action="timesheet_action_report_by_project"
-            name="By Project"
-            sequence="15"/>
-
-        <menuitem id="timesheet_menu_report_timesheet_by_task"
-            parent="menu_timesheets_reports_timesheet"
-            action="timesheet_action_report_by_task"
-            name="By Task"
-            sequence="20"/>
 
     </data>
 </odoo>
