+++ addons/hr_timesheet/views/hr_timesheet_views.xml
@@ -1,24 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
     <data>
-        <!--
-            Analytic Accounts with project
-        -->
-        <record id="account_analytic_account_view_form_inherit" model="ir.ui.view">
-            <field name="name">account.analytic.account.form.inherit</field>
-            <field name="model">account.analytic.account</field>
-            <field name="inherit_id" ref="analytic.view_account_analytic_account_form"/>
-            <field eval="18" name="priority"/>
-            <field name="arch" type="xml">
-                <xpath expr="//button[@name='toggle_active']" position="before">
-                    <button class="oe_stat_button" type="object" name="action_view_projects"
-                        icon="fa-puzzle-piece" attrs="{'invisible': [('project_count', '=', 0)]}">
-                        <field string="Projects" name="project_count" widget="statinfo"/>
-                    </button>
-                </xpath>
-            </field>
-        </record>
-
 
         <!-- Timesheet root menus -->
         <menuitem id="timesheet_menu_root"
@@ -42,9 +24,8 @@
                 <tree editable="top" string="Timesheet Activities" import="false">
                     <field name="date"/>
                     <field name="name"/>
-                    <field name="project_id" required="1"/>
-                    <field name="task_id" context="{'default_project_id': project_id}" domain="[('project_id', '=', project_id)]"/>
-                    <field name="unit_amount" widget="timesheet_uom" sum="Total"/>
+                    <field name="account_id"/>
+                    <field name="unit_amount" widget="timesheet_uom"/>
                 </tree>
             </field>
         </record>
@@ -86,8 +67,7 @@
                         <group>
                             <group>
                                 <field name="date"/>
-                                <field name="project_id" required="1"/>
-                                <field name="task_id"/>
+                                <field name="account_id"/>
                                 <field name="name"/>
                                 <field name="company_id" groups="base.group_multi_company"/>
                             </group>
@@ -123,16 +103,12 @@
                 <search string="Timesheet">
                     <field name="date"/>
                     <field name="employee_id"/>
-                    <field name="project_id"/>
-                    <field name="task_id"/>
                     <field name="name"/>
                     <filter name="mine" string="My Timesheets" domain="[('user_id', '=', uid)]"/>
                     <separator/>
                     <filter name="month" string="Date" date="date"/>
                     <group expand="0" string="Group By">
                         <filter string="Employee" name="groupby_employee" domain="[]" context="{'group_by':'employee_id'}"/>
-                        <filter string="Project" name="groupby_project" domain="[]" context="{'group_by':'project_id'}"/>
-                        <filter string="Task" name="groupby_task" domain="[]" context="{'group_by':'task_id'}"/>
                         <filter string="Timesheet Date" name="groupby_date" domain="[]" context="{'group_by':'date'}" help="Timesheet by Date"/>
                     </group>
                 </search>
@@ -146,8 +122,6 @@
                     <field name="date"/>
                     <field name="user_id"/>
                     <field name="name"/>
-                    <field name="project_id"/>
-                    <field name="task_id" context="{'default_project_id': project_id}" domain="[('project_id', '=', project_id)]"/>
                     <field name="unit_amount"/>
                     <templates>
                         <t t-name="kanban-box">
@@ -157,9 +131,6 @@
                                         <img t-att-src="kanban_image('res.users', 'image_small', record.user_id.raw_value)" t-att-title="record.user_id.value" t-att-alt="record.user_id.value" width="40" height="40" class="oe_kanban_avatar float-left"/>
                                     </div>
                                     <div class="col-10">
-                                        <div>
-                                            <strong><t t-esc="record.project_id.value"/></strong>
-                                        </div>
                                         <div class="text-muted">
                                             <span>
                                                 <t t-esc="record.name.value"/>
@@ -240,8 +211,6 @@
             <field name="context">{
                 'search_default_week':1,
                 'search_default_groupby_employee':1,
-                'search_default_groupby_project':1,
-                'search_default_groupby_task':1,
                 'default_is_timesheet': 1,
             }</field>
         </record>
