+++ addons/project_timesheet/data/project_timesheet_data.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <!-- Force Analytic account creation for projects allowing timesheet (default is True) -->
+    <function
+        model="project.project"
+        name="_init_data_analytic_account"
+        eval="[]"/>
+
+    <!-- For a perfect onboarding, we want the tour of project to be consumed
+    only by admin when installing hr_timesheet, to avoid confusing -->
+    <record id="web_tour_project_consumed_by_admin" model="web_tour.tour">
+        <field name="name">project_tour</field>
+        <field name="user_id" ref='base.user_admin'/>
+    </record>
+
+</odoo>
