+++ addons/project_timesheet/report/report_timesheet_templates.xml
@@ -67,7 +67,7 @@
             model="account.analytic.line"
             string="Timesheet Entries"
             report_type="qweb-pdf"
-            name="hr_timesheet.report_timesheet"
+            name="project_timesheet.report_timesheet"
             file="report_timesheet"
     />
 </odoo>
