+++ addons/project_timesheet/security/project_timesheet_security.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data>
+
+        <record id="timesheet_rule_portal" model="ir.rule">
+            <field name="name">Project/Task Timesheets: frontend access</field>
+            <field name="model_id" ref="analytic.model_account_analytic_line"/>
+            <field name="domain_force">[
+                '|',
+                    '&amp;',
+                        ('task_id.project_id.privacy_visibility', '=', 'portal'),
+                        ('task_id.project_id.message_partner_ids', 'child_of', [user.partner_id.commercial_partner_id.id]),
+                    '&amp;',
+                        ('task_id.project_id.privacy_visibility', '=', 'portal'),
+                        ('task_id.message_partner_ids', 'child_of', [user.partner_id.commercial_partner_id.id]),
+            ]</field>
+            <field name="groups" eval="[(4, ref('base.group_portal'))]"/>
+        </record>
+
+	</data>
+</odoo>
