+++ addons/project_timesheet/views/project_views.xml
@@ -6,11 +6,10 @@
             <field name="res_model">account.analytic.line</field>
             <field name="view_type">form</field>
             <field name="view_mode">tree,form</field>
-            <field name="view_id" ref="timesheet_view_tree_user"/>
+            <field name="view_id" ref="hr_timesheet.timesheet_view_tree_user"/>
             <field name="domain">[('project_id', '!=', False), ('is_timesheet', '=', True)]</field>
-            <field name="context">{"default_project_id": active_id, "search_default_project_id": [active_id],'default_is_timesheet': 1,
-}</field>
-            <field name="search_view_id" ref="hr_timesheet_line_search"/>
+            <field name="context">{"default_project_id": active_id, "search_default_project_id": [active_id],'default_is_timesheet': 1}</field>
+            <field name="search_view_id" ref="hr_timesheet.hr_timesheet_line_search"/>
             <field name="help" type="html">
               <p class="o_view_nocontent_smiling_face">
                 Record a new activity
@@ -207,7 +206,7 @@
             src_model="project.task"
             res_model="account.analytic.line"
             view_mode="tree"
-            view_id="hr_timesheet_line_tree"
+            view_id="hr_timesheet.hr_timesheet_line_tree"
             context="{
                 'search_default_task_id': [active_id],
                 'default_task_id': active_id,
