+++ addons/project_timesheet_holidays/__manifest__.py
@@ -13,7 +13,7 @@ Bridge module to integrate leaves in timesheet
 This module allows to automatically log timesheets when employees are
 on leaves. Project and task can be configured company-wide.
     """,
-    'depends': ['hr_timesheet', 'hr_holidays'],
+    'depends': ['project_timesheet', 'hr_holidays'],
     'data': [
         'views/res_config_settings_views.xml',
         'views/hr_holidays_views.xml',
