+++ addons/sale_timesheet/views/project_task_views.xml
@@ -31,7 +31,7 @@
     <record id="project_project_view_kanban_inherit_sale_timesheet" model="ir.ui.view">
         <field name="name">project.project.kanban.inherit.sale.timesheet</field>
         <field name="model">project.project</field>
-        <field name="inherit_id" ref="hr_timesheet.view_project_kanban_inherited"/>
+        <field name="inherit_id" ref="project_timesheet.view_project_kanban_inherited"/>
         <field name="arch" type="xml">
             <xpath expr="//a[@t-if='record.allow_timesheets.raw_value']" position="replace">
                 <a t-if="record.allow_timesheets.raw_value" name="action_view_timesheet" type="object" class="o_project_kanban_box o_project_timesheet_box" groups="project.group_project_manager">
@@ -90,7 +90,7 @@
         <record id="project_task_view_form_inherit_sale_timesheet" model="ir.ui.view">
             <field name="name">project.task.form.inherit.timesheet</field>
             <field name="model">project.task</field>
-            <field name="inherit_id" ref="hr_timesheet.view_task_form2_inherited"/>
+            <field name="inherit_id" ref="project_timesheet.view_task_form2_inherited"/>
             <field name="arch" type="xml">
                 <xpath expr="//field[@name='timesheet_ids']/tree" position="attributes">
                     <attribute name="decoration-muted">timesheet_invoice_id != False</attribute>
